#63 - refactor: S3 이미지 업로드 Service 패키지 이동
