+++ src/main/java/com/gxdxx/instagram/config/s3/S3Uploader.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.service;
+package com.gxdxx.instagram.config.s3;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
