+++ src/main/java/com/gxdxx/instagram/service/user/UserCreateService.java
@@ -5,7 +5,7 @@ import com.gxdxx.instagram.dto.response.UserSignUpResponse;
 import com.gxdxx.instagram.entity.User;
 import com.gxdxx.instagram.exception.NicknameAlreadyExistsException;
 import com.gxdxx.instagram.repository.UserRepository;
-import com.gxdxx.instagram.service.S3Uploader;
+import com.gxdxx.instagram.config.s3.S3Uploader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
